Create a small wrapper around Faraday's conn
Rename the `api` method to `conn`, since we can't modify the request
parameters directly on the initial faraday instantiation. To work around
this, we make the new `api` method take two things: a method symbol and
a block. The method symbol must match a supported method on Faraday, and
then the block is called with the request where we can eventually also
inject an authorization header.

With help and words from @alexblackie.